Reduce grid size, add click and drag
Reduced the grid container size so that the whole thing fits
inside a browser window.

Added ability to click and drag to select multiple cells.

Used ChatGPT for the latter - is this cheating? Did lots of
research, Googling and looked at examples, but couldn't find
any answers that worked well.